+++ frontend/src/components/PokemonResults/PokemonResults.tsx
@@ -2,13 +2,19 @@ import React, { FC, useContext, useEffect, useState } from "react";
 import { RouteComponentProps } from "@reach/router";
 import { PokemonContext } from "../../context/PokemonContext";
 import { MovesContext } from "../../context/MovesContext";
-import { FiltersContext } from "../../context/FiltersContext";
 import ResultsHeader from "./ResultsHeader/ResultsHeader";
 import FilterSections from "./FiltersSection/FilterSections";
 import PokemonCardList from "./PokemonCardList/PokemonCardList";
 import PokemonDetailsModal from "../PokemonDetailsModal/PokemonDetailsModal";
 import { filterVersionDataByVersion, filterVersionDataByLearnMethod } from "../../filters";
 import { Pokemon } from "../../types";
+
+type FilterState = {
+  moves: string[];
+  pokemonType: string;
+  version: string;
+  learnMethod: string;
+};
 
 const initialDisplayState = {
   id: 0,
@@ -17,14 +23,21 @@ const initialDisplayState = {
   types: [],
 };
 
+const initialFilterState: FilterState = {
+  moves: [],
+  pokemonType: null,
+  version: null,
+  learnMethod: null,
+};
+
 const PokemonResults: FC<RouteComponentProps> = (props: RouteComponentProps) => {
   const { pokemonState } = useContext(PokemonContext);
   const {
     selectedMovesState: { selectedMoves },
   } = useContext(MovesContext);
-  const { filtersState: filters, dispatch: filtersDispatch } = useContext(FiltersContext);
 
   const [displayedPokemon, setDisplayedPokmeon] = useState<Pokemon[]>([initialDisplayState]);
+  const [filters, setFilters] = useState<FilterState>(initialFilterState);
 
   useEffect(() => {
     if (process.env.NODE_ENV === "development") {
@@ -50,23 +63,18 @@ const PokemonResults: FC<RouteComponentProps> = (props: RouteComponentProps) => 
       setDisplayedPokmeon(initialDisplayData);
     }
 
-    filtersDispatch({
-      type: "CLEAR",
-      filters: {
-        moves: selectedMoves.map((selectedMove) => selectedMove.name),
-      },
+    setFilters({
+      ...initialFilterState,
+      moves: selectedMoves.map((selectedMove) => selectedMove.name),
     });
   };
 
   const applyTypeFilter = (pokemonType: string, displayText: string) => {
     const results = displayedPokemon.filter((pokemon) => pokemon.types.includes(pokemonType.toLowerCase()));
     setDisplayedPokmeon(results);
-    filtersDispatch({
-      type: "UPDATE_FILTERS",
-      filters: {
-        ...filters,
-        pokemonType: displayText,
-      },
+    setFilters({
+      ...filters,
+      pokemonType: displayText,
     });
   };
 
@@ -80,12 +88,9 @@ const PokemonResults: FC<RouteComponentProps> = (props: RouteComponentProps) => 
     filterVersionDataByVersion(results, version);
 
     setDisplayedPokmeon(results);
-    filtersDispatch({
-      type: "UPDATE_FILTERS",
-      filters: {
-        ...filters,
-        version: displayText,
-      },
+    setFilters({
+      ...filters,
+      version: displayText,
     });
   };
 
@@ -99,12 +104,9 @@ const PokemonResults: FC<RouteComponentProps> = (props: RouteComponentProps) => 
     filterVersionDataByLearnMethod(results, learnMethod);
 
     setDisplayedPokmeon(results);
-    filtersDispatch({
-      type: "UPDATE_FILTERS",
-      filters: {
-        ...filters,
-        learnMethod: displayText,
-      },
+    setFilters({
+      ...filters,
+      learnMethod: displayText,
     });
   };
