+++ frontend/src/index.tsx
@@ -4,7 +4,6 @@ import { Router } from "@reach/router";
 import { PokemonContextProvider } from "./context/PokemonContext";
 import { MovesContextProvider } from "./context/MovesContext";
 import { SearchContextProvider } from "./context/SearchContext";
-import { FiltersContextProvider } from "./context/FiltersContext";
 import { ModalDetailsContextProvider } from "./context/ModalDetailsContext";
 import Home from "./components/Home/Home";
 import PokemonResults from "./components/PokemonResults/PokemonResults";
@@ -15,14 +14,12 @@ const App: FC = () => {
       <PokemonContextProvider>
         <MovesContextProvider>
           <SearchContextProvider>
-            <FiltersContextProvider>
-              <ModalDetailsContextProvider>
-                <Router>
-                  <Home path="/" />
-                  <PokemonResults path="/results" />
-                </Router>
-              </ModalDetailsContextProvider>
-            </FiltersContextProvider>
+            <ModalDetailsContextProvider>
+              <Router>
+                <Home path="/" />
+                <PokemonResults path="/results" />
+              </Router>
+            </ModalDetailsContextProvider>
           </SearchContextProvider>
         </MovesContextProvider>
       </PokemonContextProvider>
